Rename document_store => document_stores
