+++ crates/llm-chain/src/document_stores/mod.rs
@@ -1,2 +1,2 @@
-pub mod document_store;
 pub mod in_memory_document_store;
+pub mod document_store;
